Split mapScreen into 2 Widgets - one with meta info, 1 with scaffold
